test(amqp): move out configs to stubs
