+++ packages/core/test/amqp/helpers/actions/echo.js
@@ -1,8 +1,7 @@
-const Promise = require('bluebird');
 const { ActionTransport } = require('../../../..');
 
 function EchoAction(request) {
-  return Promise.resolve(request.params);
+  return request.params;
 }
 
 EchoAction.transports = [ActionTransport.amqp, ActionTransport.internal];
